Fix path handling on windows.
os.Getwd() can return volume letter in different casings depending on the moon phase. It can be C:\ or it can be c:\. Which, in turn, ruins any checks after that, because GOPATH and other stuff is most likely C:\.

This is actually a quickfix commit - the proper fix should lowercase or uppercase EVERYTHING on Windows (because Windows paths are not case-sensitive) but it would require a lot of overhauling and will introduce additional performance penalty for Windows users. Think about it in the future.