fix vendoring if vanity URL is broken
The vendor.conf allows passing a custom URL to download the dependency,
however, the import-path may still be used to detect the repository "root"
(for versioned imports).

In situations where a vanity URL is used, and the vanity URL is broken, this
can cause downloading the dependency to fail:

This patch checks if a custom VCS URL is provided, and in that case skips the
detection of the repository root (as the VCS URL will provide this information,
and won't contain a versioned import path.

Before this change:

    vndr code.cloudfoundry.org/clock v1.0.0 https://github.com/cloudfoundry/clock.git
    Collecting initial packages
    Download dependencies
    unrecognized import path "code.cloudfoundry.org/clock" (https fetch: Get "https://code.cloudfoundry.org/clock?go-get=1": x509: certificate signed by unknown authority)

After this change:

    vndr code.cloudfoundry.org/clock v1.0.0 https://github.com/cloudfoundry/clock.git
    Collecting initial packages
    Download dependencies
        Clone https://github.com/cloudfoundry/clock.git to code.cloudfoundry.org/clock, revision v1.0.0, attempt 1/20
        Finished clone code.cloudfoundry.org/clock
    Dependencies downloaded. Download time: 2.796983086s
    Collecting all dependencies
    Clean vendor dir from unused packages
    ...

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>